requirements.txt
llama-cpp-python
huggingface-hub
ipython
black
flask
flask[async]
